Modifies the behaviour of platform preference
