+++ collect_app/src/main/AndroidManifest.xml
@@ -190,16 +190,10 @@ the specific language governing permissions and limitations under the License.
             android:name=".preferences.AboutPreferencesActivity"
             android:label="@string/app_name" />
         <activity
+            android:name=".preferences.ServerPreferencesActivity"
+            android:label="@string/app_name" />
+        <activity
             android:name=".preferences.PreferencesActivity"
-            android:label="@string/app_name" />
-        <activity
-            android:name=".preferences.GooglePreferencesActivity"
-            android:label="@string/app_name" />
-        <activity
-            android:name=".preferences.OtherPreferencesActivity"
-            android:label="@string/app_name" />
-        <activity
-            android:name=".preferences.AggregatePreferencesActivity"
             android:label="@string/app_name" />
         <activity
             android:name=".preferences.FormMetadataActivity"
