+++ collect_app/src/main/java/org/odk/collect/android/preferences/AboutPreferencesActivity.java
@@ -26,7 +26,10 @@ public class AboutPreferencesActivity extends PreferenceActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        getFragmentManager().beginTransaction().replace(android.R.id.content, new AboutPreferencesFragment()).commit();
+        getFragmentManager()
+                .beginTransaction()
+                .replace(android.R.id.content, new AboutPreferencesFragment())
+                .commit();
         setTitle(getString(R.string.about_preferences));
     }
 }
