--- collect_app/src/main/java/org/odk/collect/android/preferences/OtherPreferencesActivity.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (C) 2014 Nafundi
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package org.odk.collect.android.preferences;
-
-import android.os.Bundle;
-
-/**
- * Handles 'other' specific preferences.
- *
- * @author Carl Hartung (chartung@nafundi.com)
- */
-public class OtherPreferencesActivity extends PreferencesActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        getFragmentManager()
-                .beginTransaction()
-                .replace(android.R.id.content, new OtherPreferencesFragment())
-                .commit();
-    }
-}
