+++ collect_app/src/main/res/xml/aggregate_preferences.xml
@@ -23,5 +23,5 @@
             android:key="password"
             android:title="@string/password" />
     </PreferenceCategory>
-   
+
 </PreferenceScreen>
