+++ collect_app/src/main/res/xml/other_preferences.xml
@@ -1,4 +1,5 @@
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+
     <PreferenceCategory
         android:key="@string/other_preferences"
         android:title="@string/other_preferences">
