+++ collect_app/src/main/res/xml/server_preferences.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+    <ListPreference
+        android:id="@+id/protocol"
+        android:defaultValue="odk_default"
+        android:dialogTitle="@string/change_protocol"
+        android:entries="@array/protocol_entries"
+        android:entryValues="@array/protocol_entry_values"
+        android:key="protocol"
+        android:title="@string/type" />
+</PreferenceScreen>
